+++ app/models/Pedido.php
@@ -6,18 +6,10 @@ use DateTime;
 
 class Pedido extends Model
 {
-    private int $CD_PEDIDO;
-    private DateTime $DT_PEDIDO;
-    private string $DS_STATUS;
-    private float $VL_TOTAL;
-    private float $VL_DESCONTO;
-    private float $VL_LIQUIDO;
-    private int $NU_PARCELAS;
-    private string $DS_TIPO;
-    private int $CD_USUARIO;
-    private int $CD_CLIENTE;
     private int $CD_ORCAMENTO;
-    private int $CD_VEICULO;
+    private DateTime $DT_ORCAMENTO;
+    private float $VL_ORCAMENTO;
+    private int $CD_CLIENTE;
 
     public function __construct()
     {
@@ -27,104 +19,37 @@ class Pedido extends Model
     //codigo
     public function getCodigo(): int
     {
-        return $this->CD_PEDIDO;
+        return $this->CD_ORCAMENTO;
     }
 
     public function setCodigo(int $codigo): void
     {
-        $this->CD_PEDIDO = $codigo;
+        $this->CD_ORCAMENTO = $codigo;
     }
 
-    //Data do pedido
+    //Data do orçamento
     public function getData(): DateTime
     {
-        return $this->DT_PEDIDO;
+        return $this->DT_ORCAMENTO;
     }
 
     public function setData(string $data): void
     {
-        $this->DT_PEDIDO = $data;
+        $this->DT_ORCAMENTO = $data;
     }
 
-    //Status
-    public function getStatus(): string
-    {
-        return $this->DS_STATUS;
-    }
-
-    public function setStatus(string $status): void
-    {
-        $this->DS_STATUS= $status;
-    }
-
-    //Valor total do pedido
+    //Valor total do orçamento
     public function getValor(): float
     {
-        return $this->VL_TOTAL;
+        return $this->VL_ORCAMENTO;
     }
 
     public function setValor(float $valor) : void 
     {
-        $this->VL_TOTAL= $valor;   
+        $this->VL_ORCAMENTO= $valor;   
     }
 
-    //Valor do desconto pedido
-    public function getDesconto(): float
-    {
-        return $this->VL_DESCONTO;
-    }
-
-    public function setDesconto(float $desconto) : void 
-    {
-        $this->VL_DESCONTO = $desconto;   
-    }
-
-    //Valor liquido do pedido
-    public function getValorLiquido(): float
-    {
-        return $this->VL_LIQUIDO;
-    }
-
-    public function setValorLiquido(float $valorLiquido) : void 
-    {
-        $this->VL_LIQUIDO = $valorLiquido;   
-    }
-
-    //Numero de parcelas do pedido
-    public function setParcelas(int $parcelas) : void
-    {
-        $this->NU_PARCELAS = $parcelas;
-    }
-
-    
-    public function getParcelas() : int
-    {
-        return $this->NU_PARCELAS;
-    }
-
-    //Tipo de pagamento
-    public function getTipoPagamento() : string 
-    {
-        return $this->DS_TIPO;    
-    }
-
-    public function setTipoPagamento(string $tipo) : void 
-    {
-        $this->DS_TIPO = $tipo;
-    }
-
-    //Usuario que cadastrou o pedido
-    public function getUsuario() : int 
-    {
-        return $this->CD_USUARIO;    
-    }
-
-    public function setUsuario(Usuario $usuario) : void 
-    {
-        $this->CD_USUARIO = $usuario->getCodigo();    
-    }
-
-    //Cliente que fez o pedido
+    //Cliente que fez o orçamento
     public function getCliente() : int 
     {
         return $this->CD_CLIENTE;    
@@ -133,16 +58,5 @@ class Pedido extends Model
     public function setCliente(Cliente $cliente) : void 
     {
         $this->CD_CLIENTE = $cliente->getCodigo();    
-    }
-
-    //Veiculo que sera feito o pedido
-    public function getVeiculo() : int 
-    {
-        return $this->CD_VEICULO;    
-    }
-
-    public function setVeiculo(Veiculo $veiculo) : void 
-    {
-        $this->CD_VEICULO = $veiculo->getCodigo();    
     }
 }
